requirements.txt
transformers==4.36.2
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
numpy==1.24.4
scipy==1.10.1
scikit-image==0.21.0
opencv-python==4.7.0.72
pillow==9.4.0
diffusers==0.25.0
transformers==4.36.2
accelerate==0.26.1
matplotlib==3.7.4
tqdm==4.64.1
config==0.5.1
einops==0.7.0
onnxruntime==1.16.2
kafka-python==2.0.2
basicsr
av
fvcore
cloudpickle
omegaconf
pycocotools
gradio
kafka-python
imagekitio
requests-toolbelt
config==0.5.1
einops==0.7.0
ninja==1.10.2
basicsr
urllib3==1.26.15
